优化

+++ tests/Database/QueryMonitor/TestCase.php
@@ -10,7 +10,7 @@ trait TestCase
     protected function getQueryExecuted(
         ?string $sql = null,
         ?array $bindings = null,
-        int $time = 1000
+        float $time = 1000
     ): QueryExecuted {
         return new QueryExecuted(
             $sql ?: 'select username from users where uid = ?',
